requirements.txt
asgiref==3.5.2
Django~=4.0.5
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2

django-filter~=22.1
pip~=22.0.2
wheel~=0.37.1
cryptography~=3.4.8
Pillow~=9.0.1
keyring~=23.5.0
colorama~=0.4.4
PyYAML~=5.4.1
setuptools~=59.6.0
future~=0.18.2
requests~=2.25.1